Fixed goodmanga pagination, bato.to chapter list.
Changed the way chapters_pagination_hidden works, now I use it to
find which pagination links to skip.
Because bato.to permits multiple translations of chapter in the same
language, script now grabs first one, and skips others.
Chapter links are now always sorted so that they're in ascending order.